+++ api/api.go
@@ -3,8 +3,8 @@ package api
 import (
 	"github.com/gin-gonic/gin"
 	cronjobs "se-school-case/pkg/domain/cron-jobs"
-	"se-school-case/pkg/domain/rate"
-	"se-school-case/pkg/domain/subscriber"
+	"se-school-case/pkg/domain/rates"
+	"se-school-case/pkg/domain/subscribers"
 	"se-school-case/pkg/initializer"
 )
 
@@ -20,8 +20,8 @@ func NewApi() Api {
 	engine := gin.Default()
 	deps := wireDependencies()
 	deps.cronService.StartScheduler()
-	rate.NewHandler(deps.rateService).Register(engine)
-	subscriber.NewHandler(deps.subscriberService).Register(engine)
+	rates.NewHandler(deps.rateService).Register(engine)
+	subscribers.NewHandler(deps.subscriberService).Register(engine)
 	cronjobs.NewHandler(deps.mailService).Register(engine)
 	return &api{engine}
 }
