+++ api/dependencies.go
@@ -3,27 +3,27 @@ package api
 import (
 	"gorm.io/gorm"
 	cron_jobs "se-school-case/pkg/domain/cron-jobs"
-	"se-school-case/pkg/domain/mail"
-	"se-school-case/pkg/domain/rate"
-	"se-school-case/pkg/domain/subscriber"
+	"se-school-case/pkg/domain/mails"
+	"se-school-case/pkg/domain/rates"
+	"se-school-case/pkg/domain/subscribers"
 	"se-school-case/pkg/initializer"
 	"se-school-case/pkg/util/constants"
 )
 
 type dependencies struct {
-	subscriberService subscriber.SubscriberInterface
-	rateService       rate.RateInterface
-	mailService       mail.MailInterface
+	subscriberService subscribers.SubscriberInterface
+	rateService       rates.RateInterface
+	mailService       mails.MailInterface
 	cronService       cron_jobs.CronJobsInterface
 }
 
 func wireDependencies() *dependencies {
 	initEnv()
 	db := connectToDb()
-	fetchService := rate.NewRateFetchService()
-	rateService := rate.NewService(db, &fetchService)
-	subscriberService := subscriber.NewService(db)
-	mailService := mail.NewService(&subscriberService, &rateService)
+	fetchService := rates.NewRateFetchService()
+	rateService := rates.NewService(db, &fetchService)
+	subscriberService := subscribers.NewService(db)
+	mailService := mails.NewService(&subscriberService, &rateService)
 	cronService := cron_jobs.NewService(&mailService)
 	return &dependencies{
 		subscriberService: &subscriberService,
