+++ pkg/domain/cron-jobs/handler.go
@@ -26,12 +26,12 @@ func (h *CronJobsHandler) Register(engine *gin.Engine) {
 // swagger:route POST /api/notify CronJobs postExplicitlyNotify
 // Explicitly notify all subscribers
 //
-// Sends an email notification to all subscribers about the exchange rate.
+// Sends an email notification to all subscribers about the exchange rates.
 //
 // responses:
 //
 //	200: body:gin.H{"message": "Successfully notified all users."}
 func (h *CronJobsHandler) PostExplicitlyNotify(context *gin.Context) {
-	h.mailService.SendEmailToAll("Exchange rate notification", constants.TEMPLATE_PATH)
+	h.mailService.SendEmailToAll("Exchange rates notification", constants.TEMPLATE_PATH)
 	context.JSON(http.StatusOK, gin.H{"message": "Successfully notified all users."})
 }
