+++ pkg/domain/rates/service.go
@@ -1,4 +1,4 @@
-package rate
+package rates
 
 import (
 	"errors"
@@ -18,13 +18,13 @@ func NewService(repository *gorm.DB, rateFetch RateFetchInterface) RateService {
 	return RateService{repository: repository, fetchService: rateFetch}
 }
 
-var ErrNoRateFound = errors.New("no rate found")
+var ErrNoRateFound = errors.New("no rates found")
 
 func (s *RateService) GetRate() (model.Rate, error) {
 	rate, err := s.getLatestRate()
 	if err != nil {
 		if errors.Is(err, ErrNoRateFound) || err.Error() == "record not found" {
-			// Fetch exchange rate if no rate found
+			// Fetch exchange rates if no rates found
 			exchangeRate, fetchErr := s.fetchService.FetchExchangeRate()
 			if fetchErr != nil {
 				return model.Rate{}, fetchErr
@@ -39,7 +39,7 @@ func (s *RateService) GetRate() (model.Rate, error) {
 		}
 	}
 
-	// Check if the rate is more than 1 hour old
+	// Check if the rates is more than 1 hour old
 	if time.Since(rate.CreatedAt) > constants.UpdateInterval {
 		exchangeRate, fetchErr := s.fetchService.FetchExchangeRate()
 		if fetchErr != nil {
@@ -56,7 +56,7 @@ func (s *RateService) GetRate() (model.Rate, error) {
 }
 
 func (s *RateService) SaveRate(currencyFrom string, currencyTo string, exchangeRate float64) {
-	// Delete existing rate records where CurrencyFrom and CurrencyTo match
+	// Delete existing rates records where CurrencyFrom and CurrencyTo match
 	if err := s.repository.Where("currency_from = ? AND currency_to = ?",
 		currencyFrom, currencyTo).Delete(&model.Rate{}).Error; err != nil {
 		log.Printf("Error deleting old exchange rates: %v", err)
@@ -65,7 +65,7 @@ func (s *RateService) SaveRate(currencyFrom string, currencyTo string, exchangeR
 
 	rate := model.Rate{CurrencyFrom: currencyFrom, CurrencyTo: currencyTo, Rate: exchangeRate}
 	if err := s.repository.Create(&rate).Error; err != nil {
-		log.Printf("Error writing exchange rate to database: %v", err)
+		log.Printf("Error writing exchange rates to database: %v", err)
 		return
 	}
 }
