+++ pkg/domain/subscribers/email-dto.go
@@ -1,4 +1,4 @@
-package subscriber
+package subscribers
 
 type EmailDto struct {
 	Email string `form:"email" binding:"required,email"`
