+++ pkg/domain/subscribers/service.go
@@ -1,4 +1,4 @@
-package subscriber
+package subscribers
 
 import (
 	"errors"
